Refactor reach-and-frequency metric (#1341)
In Report V2 the `FrequencyHistogramParams` will be renamed to
`ReachAndFrequencyParams`. Note the `FrequencyHistogramParams` already
includes `ReachParams`. Setting this new Params structure cause a Reach
and Frequency Results to be created. As such, metric result for
`ReachAndFrequency` will be updated to replace `HistogramResult` with a
`ReachAndFrequencyResult` that includes a `HistogramResult` and a
`ReachResult` as sub-messages.

Note:
* This PR changes an existing Liquibase changeset, meaning the DB will
need to be wiped.
* This PR renames a field in `MetricSpecConfig`, so anyone with an
existing deployment will need to update their config.

Report V1 will not be updated.