Add a metric reuse mechanism based on timed ReportingMetricEntries (#1347)
Given the fact that `MetricCalculationSpec` is now an individual
resource and can be reused, we can utilize this to identify a set of
metrics across reports. Note that the granularity is "a set of metrics"
that is associated with `MetricCalculationSpec`.

When a `MetricCalculationSpec` is reused in a different report, by
identifying the pair of `MetricCalculationSpec` and the `ReportingSet`
and the `TimeInterval`s from the report, we are able to identify the
existing set of metrics created from the previous reports. This could be
done during the internal report creation by the following:
1. Left join the table `Metrics` and
`MetricCalculationSpecReportingMetrics` by `MeasurementConsumerId` and
`MetricId`.
2. Find the set of metrics given `MeasurementConsumerId`,
`ReportingSetId`, `MetricCalculationSpecId`, a set of
`TimeIntervalStart` and `TimeIntervalEndExclusive`.
3. Repeat the process for all `ReportingMetricEntries`, i.e. pairs of
`MetricCalculationSpecs` and `ReportingSets`.

At the step (2) above, the reason why we could get a set of metrics is
because a `MetricCalculationSpec` could correspond to multiple metrics
due to the fan-out of metric types, groupings, and time intervals.