Limit updating the status of a task on attempt completion
Specifically: if attempt has caused a temporary failure, do not update
the status of the task to READY_TO_RETRY if the attempt being marked as
failed is not the latest attempt.

Justification: We have noticed that there are cases where multiple
IN_PROGRESS attempts exist for a single task. Whether those are active
or not is unclear, but the result of this was if any fail or time out
the task is set to READY_TO_RETRY while other attempts are currently
IN_PROGRESS (which is not desirable). This way, if there is somehow a
newer attempt that was created, we trust that attempt will update the
task state when it completes or times out, rather than allowing the task
to issue another attempt before newer attempts complete.