Introduce the concept of known EventGroup metadata types. (#1512)
Known EventGroup metadata types are types expected to be known by all callers. These need not be included in the EventGroupMetadataDescriptor FileDescriptorSet, and are instead loaded separately.

This also enforces uniqueness for EventGroupMetadataDescriptor message types. Note that this only affects the creation of new resources. Existing resources may exist that already violate this constraint.

Includes a workaround for bazelbuild/rules_proto#203.

Closes #1511