Prepare example Kingdom-less exchange daemon for GKE. (#1698)
Summary:
* Creates `.cue` files for deploying Kingdom-less daemon to GKE.
* Adds Kingdom-less daemon flags.
* Uses `@ArgGroup` to ensure mutual exclusitivity between Kingdom-based
flags and Kingdom-less flags.
* Updates the example daemon main class to build the correct type of
`ApiClient` and `CertificateManager` based on which group of flags was
set.
* Several other tweaks to make the deployment work:
* Factory function for building a shared storage client needed to be
`suspend`ing.
* Explicitly call `withDefaultCredentails()` when building
`GcpKmsClient`. This used to happen by default, but seems to have been
(inadvertently?) removed in a previous PR.
    * In `KingdomlessApiClient.kt`:
* Make sure that `ExchangeWorkflowDependencyGraph` is built in step
order. The dependency graph will throw if e.g. you try to mark a step as
completed before its prerequisite steps.
* Add retries when loading checkpoints from shared storage, as a
temporary measure to mitigate race conditions. A longer term solution
will be to add a direct upload option to `StorageClient`, but need to
investigate this further.