fix!: Ensure default network policies are created (#1768)
The previous behavior was to add some default ingresses and egresses to every NetworkPolicy other than default-deny. The flaw with this behavior is that if no NetworkPolicy matches the pod, the pod will not have these default ingresses and egresses. This instead creates separate NetworkPolicy objects for defaults.

This also fixes an issue where the opentelemetry-collector NetworkPolicy was missing its pod selector and was therefore applying to all pods.