feat: Extend last report time interval to end of report (#1869)
feat: The last report time interval covers the end of the report

This adds a time interval that covers the end of the report even if the
look back window goes past the report_end. This additional time interval
just cuts off at report_end. Currently, if the look back window goes
past the report_end, there is no time interval that partially covers it.