refactor: Extract listResources utility function for handling pagination (#1923)
This also fixes an issue where EventGroupsServiceTest was waiting for a
real 30s deadline to pass.
Also fixes the simulator issue of PR
https://github.com/world-federation-of-advertisers/cross-media-measurement/pull/1927:
The MC and EDP simulators were only reading the first page of results
from ListEventGroups. This means that in environments with more
EventGroups, not all test EventGroups would be read. This is exacerbated
by
https://github.com/world-federation-of-advertisers/cross-media-measurement/pull/1916,
as the default page size is now smaller.