feat: Add flag for specifying the the VID model line (#1931)
Support VID Model Line rollout using Reporting Server flags. Two flags
are added:

–default-vid-model-line=<String>
Specifies the VID model line value to be used for the “model_line” field
in MeasurementSpec This in turn determines what VID model line that EDPs
will use to fulfill Requisitions. If empty EDPs must agree offline what
the default is. Once enabled we encourage that this case become an
error.

–measurement-consumer-vid-model-line=Map<String, String>
A map of MC names to model lines. When a Metric is created, if the
Metric is for an MC present in this map, then the model line provided as
the value will be used for the “model_line” field in MeasurementSpec.