perf: Avoid full table scan of Measurements in Requisitions query (#2041)
From testing on halo-cmm-dev, we see that ListRequisitions filtering by Measurement State and Requisition State results in a full table scan of the Measurements table. This is the exact call used by EDP simulators once per second.

In addition to making the query itself slow, this can result in locking all Measurement rows. This in turn can result in long lock waits for writes to the Measurements table.