feat: Adds function that includes attributes with PopulationSpec when converting from SyntheticPopulationSpec (#2053)
Add a function to PopulationSpecConverter that returns a PopulationSpec
with the attributes field populated. Modifying the existing function
broke the EDP simulator when using other event templates (i.e. not Halo
test templates), so this PR creates a new function that is independent
from the original. It will be used later in the KIND test for
PopulationRequisitionFulfiller.

---------

Co-authored-by: jojijac0b <jojijac0b@github.com>