chore: Avoid changing individual bytes in SecretData (#2078)
This prepares for Tink 3.0.

SecretData will become a container which can only be manipulated in
Bulk. See
https://github.com/tink-crypto/tink-cc/blob/main/tink/util/secret_data_internal_class.h
for the API.

Unfortunately it is a little bit difficult for me as Tink maintainer to
deprecate these functions since currently SecretData is an alias for
std::vector<uint8_t, SecretAllocator>. This will however change.