fix: Use approximated variance for measurements derived from the unique reach ones. (#2131)
In the report post-processing program, when parsing unique reach
measurements, e.g. reach(X - Y), sometimes we need to infer the
measurement reach(Y) from reach(X - Y) and reach (X union Y). This
involves computing the variance of the measurement reach(Y).

Previously, it was assumed that reach(X - Y) is computed directly from
reach(Y) and reach(X union Y), however, this does not seem to be true
(e.g. it is computed from primitive regions instead).

A temporary solution is to use the maximum of the variance of reach(X -
Y) and reach(X union Y) for reach(Y). This is expected to not have much
effect on the output of the report processor program.