refactor: Round noise correction output to nearest integer. (#2132)
The output of the quadratic solver is a vector of float numbers and they
need to be converted to integers. Currently, the conversion is done by
casting, e.g. value -> int(value), which is equivalent to truncating the
value. It's preferable to use round(value) to round it to the nearest
integer.