feat!: Use Access API for Reporting service authorization (#2020)
BREAKING CHANGE: The proxied `DataProviders` and `EventGroupMetadataDescriptors` services use a fixed MC's credentials to call the Kingdom. This can be controlled by the `--system-measurement-consumer` option to the Reporting public API server.
BREAKING CHANGE: `MeasurementConsumerConfig` has a new `offline_principal` field. This must be specified in order for Report scheduling to occur for a given MeasurementConsumer.