feat!: Persist EventGroup data availability interval in DB columns (#2216)
This moves persistence of EventGroup.data_availability_interval from EventGroupDetails to DB columns of type TIMESTAMP. Doing so allows for comparison operators in queries, which is necessary for #2124. This results in data loss of existing data availability values, which should be fine as the field is not yet being used.

BREAKING CHANGE: Any existing EventGroup data availability intervals will be lost