fix: remove event group sync iam binding (#2429)
The Event Group Sync Cloud Function is deployed via the
`deploy_edp_aggregator_cloud_functions` workflow using a gcloud command.

Terraform was previously attempting to assign the `roles/run.invoker
`permission to the Data Watcher service account for this function, but
failed because the function does not yet exist at the time of the
Terraform run.

This PR removes the IAM binding from Terraform to prevent the failure.
The binding will be applied separately using a gcloud command in a
follow-up PR.

---------

Co-authored-by: marcopremier <marcopremier@meta.com>