fix: Increase memory limit of computation control server deployment (#2558)
OOMKilled events were observed when running the correctness test in Halo-managed environments. The increase appears to be from classes, as there's a jump in Metaspace usage starting with nightly/20250626.1.

RELNOTES: Operators will likely want to increase the memory limit of the Duchy computation control service to at least 576 MiB to accommodate dependency updates.