fix: Ensure ListBasicReports PageToken always contains create_time_after filter if set (#2815)
If a ListBasicReportsRequest contains a create_time_after filter and the
response contains a page token. Using the page token will result in the
response returning a page token without the create_time_after filter,
which results in a request/page token mismatch if it is used. This is
now fixed.

Issue: #2816