fix(report-processor): always write report post processor log to bucket. (#2885)
Currently, when the python level noise correction code throws an
exception, the Kotlin terminates without logging the report post
processor output to the storage bucket. With this change, when python
level code throws an exception, the report post processor log will
record an INTERNAL_ERROR issue, and will be written to the storage
bucket.

This PR closes #2884.

Issue: #2884