feat: Create Job to check Reports backing BasicReports (#2881)
BasicReport results are determined by Report results. This job checks
the Report for every BasicReport in the state REPORT_CREATED. If the
Report is in the state FAILED, the BasicReport state is set to FAILED.
If the Report is in the state SUCCEEDED, nothing will happen right now,
but will in a future change.

Issue: #2606