fix: report_result and corresponding service (#2973)
The filter component, without the IQF, is a list of EventFilter so
event_filter is changed to a repeated event_filters. ReportResultValues
is unused so report_result_values is added. ReportResultValues is also
updated to include NoisyMetricSet, which includes the frequency
histogram and variance info. reporting_window_result and
reporting_set_result are renamed to be plural. It seems like it would be
complicated to add cumulative and noncumulative noisy results separately
in Spanner and it is fairly doable in code so the methods are merged
back into one.

Issue: #2972