fix(report-processor): handle report with non-video EDP. (#3117)
In the function that computes the CI Union checks, a list of EDP
combinations is generated and the check is done for each of them.
However, when an EDP does not support MRC, one EDP combination exists
for AMI, but not MRC. When we perform the check for that EDP combination
on MRC, the corresponding measurement does not exist and causes the
problem.

This PR adds a check if an EDP combination exists for an impression
filter before running the union check.

This PR closes
https://github.com/world-federation-of-advertisers/cross-media-measurement/issues/2880.