Fix: bug in the RequisitionFetcher where existing requisition metadata with state QUEUED or PROCESSING where treated as new (#3121)
# Summary

Treat QUEUED and PROCESSING requisition metadata states as existing
instead of new

# Description

This change resolves a bug in RequisitionFetcher where existing
requisition metadata in state QUEUED or PROCESSING was incorrectly
treated as new. As a result, the system attempted to re-group and
recreate metadata for requisitions already in progress, leading to a
gRPC ALREADY_EXISTS exception.

# What’s Fixed

QUEUED and PROCESSING metadata entries are now recognized as existing,
same as STORED.

# Tests

Added a test to ensure existing metadata in QUEUED and PROCESSING states
prevents creation of a new grouped requisition

Issue: #3135 
Closes: #3135

---------

Co-authored-by: marcopremier <marcopremier@meta.com>