fix(report-processor): use double type for frequency histogram value. (#3180)
Currently, the type of frequency histogram bin value is double. However,
report processor assumes that these values are integers. The assignment
from report proto to report summary proto has an implicit conversion
from double to Long, which is equivalent to a rounding down operation.
Due to the rounding down issue, the reach becomes larger than the sum of
the frequency counts, and the report processor tends to correct it. The
problem appears when the EDP doesn't add noises to the measurements,
thus, the change to the reach could be flagged as "large correction". To
avoid this issue, the report processor now supports the double type for
frequency histogram bin values.

Closes #3159

Issue: #3159