refactor: Refer to processed/unprocessed instead of noisy/de-noised
In the Reporting internal API, the terms "noisy" and "de-noised" have been used incorrectly. The Report Processor does not de-noise, but rather processes Report results to add derived metrics and ensure constraints are met. Similarly in the case where differential privacy is not used, the unprocessed results aren't technically "noisy".

This updates references within that API. To avoid churn, it avoids any cases that would require a schema update.